15122017AIS

+++ OrderManagement/Entity/vwOrderDetail.cs
@@ -27,14 +27,14 @@ namespace OrderManagement.Entity
         public string OrderDay { get; set; }
         public string ProductName { get; set; }
         public string ProductAbbr { get; set; }
-        public Nullable<byte> Category { get; set; }
+        public Nullable<int> Category { get; set; }
         public Nullable<decimal> Price { get; set; }
         public Nullable<int> Unit { get; set; }
         public Nullable<int> Amount { get; set; }
         public string CitizenID { get; set; }
         public string CustomerName { get; set; }
         public string CustomerAbbr { get; set; }
-        public Nullable<byte> Zone { get; set; }
+        public Nullable<int> Zone { get; set; }
         public string Address1 { get; set; }
         public string Address2 { get; set; }
         public string Phone { get; set; }
